backend/requirements.txt
fastapi
sqlalchemy
python-dotenv
requests
python-jose[cryptography]
pydantic
uvicorn
svix